UNIFICACION FORMULARIO CON EL BACKEND Y implementacion de n8n

+++ pqrs_back/src/main/java/com/ideapro/pqrs_back/pqrs/service/PqrsService.java
@@ -3,7 +3,7 @@ package com.ideapro.pqrs_back.pqrs.service;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Random;
+import java.util.concurrent.ThreadLocalRandom;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,20 +11,29 @@ import org.springframework.stereotype.Service;
 import com.ideapro.pqrs_back.pqrs.model.Pqrs;
 import com.ideapro.pqrs_back.pqrs.repository.PqrsRepository;
 
+import jakarta.persistence.EntityNotFoundException;
+
 @Service
 public class PqrsService {
+
     @Autowired
     private PqrsRepository pqrsRepository;
 
     public Pqrs crearPqrs(Pqrs pqrs) {
+        // Fecha automática
         pqrs.setFechaRegistro(LocalDateTime.now());
-        pqrs.setNumeroRadicado(generarRadicado()); // Genera radicado con la función
+
+        // Genera radicado
+        pqrs.setNumeroRadicado(generarRadicado());
+
         return pqrsRepository.save(pqrs);
     }
 
-    public String generarRadicado() {
+    private String generarRadicado() {
         String prefijo = "PQRS-NUM-";
-        String consecutivo = String.format("%07d", new Random().nextInt(9999999));
+        String consecutivo = String.format("%07d",
+                ThreadLocalRandom.current().nextInt(1, 9_999_999)
+        );
         String anio = String.valueOf(LocalDate.now().getYear());
         return prefijo + consecutivo + "-" + anio;
     }
@@ -34,15 +43,18 @@ public class PqrsService {
     }
 
     public Pqrs obtenerPqrs(Long id) {
-        return pqrsRepository.findById(id).orElse(null);
+        return pqrsRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("PQRS no encontrada con id: " + id));
     }
 
     public void eliminarPqrs(Long id) {
+        if (!pqrsRepository.existsById(id)) {
+            throw new EntityNotFoundException("No se puede eliminar. PQRS no encontrada con id: " + id);
+        }
         pqrsRepository.deleteById(id);
     }
 
     public List<Pqrs> buscarPorNumeroRadicado(String numeroRadicado) {
         return pqrsRepository.findByNumeroRadicado(numeroRadicado);
     }
-
 }
